move ruby extraction rules behind the indexer language registry
S3-RUBY (ARCH-012), the sibling of S3-KOTLIN (#1676) and one rollback unit on
its own. Ruby's tree-sitter rule file, compiled-rule cache, member-call
callsite marker, and receiver-call resolution engine move out of the indexer
god file into `codestory-indexer/src/languages/ruby.rs`, behind the existing
`LanguageExtraction` registry in `languages/mod.rs`. Every language-keyed
dispatch that spelled "ruby" now reads the registry first and keeps its
residual `match` for the languages that have not moved; the last sibling
package deletes those arms.

The move is mechanical: 15 functions leave `lib.rs` and land in
`languages/ruby.rs` alongside one new `ruby_language()` parser constructor for
the registry row. Diffing the deleted `lib.rs` lines against the moved bodies,
ignoring blank lines, leaves only the two declared renames
(`collect_ruby_receiver_call_edges` -> `receiver_call_specs`,
`ruby_receiver_call` -> `member_call`, the latter `pub(crate)` because `lib.rs`
still calls it) plus the nine registry-shaped edits listed below.

Four Ruby surfaces stay in `lib.rs` on purpose, because each is a shared seam
rather than Ruby content: `collect_ruby_bare_call_edges` and
`collect_ruby_runtime_import_specs` (per-language dispatches with non-uniform
signatures across all sixteen languages), `annotate_ruby_member_call_placeholders`
(it reaches into the edge buffer, the dedup key set, and the index feature
flags, none of which the registry row describes), and `collect_rails_route`
(the same per-framework seam Kotlin's Ktor collector left behind).
`quoted_literal_surface` also stays because the runtime-import seam still calls
it. `LanguageRuleset::Ruby` stays because the enum keys every language's
compiled-rule cache at once.

Ruby's callsite marker is not emitted by the rule file: `rules/ruby.scm` carries
no `call_syntax` attribute, so the row pairs `member_callsite_marker` and
`graph_call_syntax` as `None`/`None` and the pre-pass keeps stamping
`languages::ruby::MEMBER_CALLSITE_MARKER` directly.

Proof: `tests/language_extraction_snapshot.rs` gains one `SnapshotCase` row and
two goldens pinning the full rendered projection of both Ruby fixtures — the
fidelity_lab file through the store pipeline (resolution included) and the
tictactoe file through raw `index_file`. Both goldens were captured before the
production edit and re-captured after it; the diff is empty and the sha256s are
unchanged (7d655e24.. / ac3aa138..). Kotlin's goldens are also untouched
(a26ee4c6.. / 79df1ceb..).

The threshold suites are not sufficient on their own. Changing Ruby's
qualified-name delimiter from "." to "::" leaves both `fidelity_regression`
(8 passed) and `tictactoe_language_coverage` (12 passed) green while both new
snapshots go red on `ArtificialPlayer::minMax` vs `ArtificialPlayer.minMax`.
Dropping the receiver-call engine (`receiver_call_specs: None`) is caught by
both the snapshot and `fidelity_regression`: the resolved
`orchestrate_ruby -> Workflow.run` CALL edge degrades to an unresolved
`UNKNOWN:run` placeholder.

Declared behavior changes:
- None to any indexing, resolution, or rendering output. Ruby
  fidelity_regression and tictactoe_language_coverage projections are proven
  byte-identical before and after, and packet_search_eval is unmoved.
- `LanguageRuleset::Ruby`'s arm in `compiled_rules` is now unreachable and
  returns a typed error instead of compiled rules; the registry answers first.
  Fail-closed by construction, exactly as the Kotlin arm already is.
- One additive row in `codestory_contracts::language_support::LANGUAGE_COMMENT_PROFILES`
  (`ruby` -> `#`), which is the same marker `codestory-cli`'s roster held before
  the move; the CLI roster drops its `"ruby"` entry and reads the registry.
  `LANGUAGE_SUPPORT_PROFILES` is unchanged, so extension routing, claim tiers,
  `supported_extensions`, and `is_structural_source_path` are identical.
- The CLI comment roster and the indexer route-comment roster still disagree
  where they disagreed before; neither was unified.
- No version surface touched; everything stays 0.16.3.

Closes #1686
Refs #1243
Refs #1179
