+++ crates/codestory-indexer/src/semantic/mod.rs
@@ -597,7 +597,6 @@ fn language_family_bucket(language: &'static str) -> &'static str {
         "rust" => "rust",
         "java" => "java",
         "go" => "go",
-        "ruby" => "ruby",
         "php" => "php",
         "csharp" => "csharp",
         "swift" => "swift",
