+++ crates/codestory-indexer/tests/fixtures/language_snapshots/ruby_fidelity.txt
@@ -1,0 +1,46 @@
+node CLASS name=ConsoleNotifier qn=ConsoleNotifier canonical=<ROOT>/fidelity.rb:ConsoleNotifier line=9 col=1 end=13:4 file=FILE:<ROOT>/fidelity.rb@1
+node CLASS name=Notifier qn=Notifier canonical=<ROOT>/fidelity.rb:Notifier line=3 col=1 end=7:4 file=FILE:<ROOT>/fidelity.rb@1
+node CLASS name=Repository qn=Repository canonical=<ROOT>/fidelity.rb:Repository line=15 col=1 end=19:4 file=FILE:<ROOT>/fidelity.rb@1
+node CLASS name=Workflow qn=Workflow canonical=<ROOT>/fidelity.rb:Workflow line=23 col=1 end=38:4 file=FILE:<ROOT>/fidelity.rb@1
+node FILE name=<ROOT>/fidelity.rb qn=<ROOT>/fidelity.rb canonical=<ROOT>/fidelity.rb:<ROOT>/fidelity.rb:1 line=1 col=1 end=42:0 file=FILE:<ROOT>/fidelity.rb@1
+node METHOD name=ConsoleNotifier.notify qn=ConsoleNotifier.notify canonical=<ROOT>/fidelity.rb:ConsoleNotifier.notify#0 line=10 col=3 end=12:6 file=FILE:<ROOT>/fidelity.rb@1
+node METHOD name=Notifier.notify qn=Notifier.notify canonical=<ROOT>/fidelity.rb:Notifier.notify#0 line=4 col=3 end=6:6 file=FILE:<ROOT>/fidelity.rb@1
+node METHOD name=Repository.save qn=Repository.save canonical=<ROOT>/fidelity.rb:Repository.save#0 line=16 col=3 end=18:6 file=FILE:<ROOT>/fidelity.rb@1
+node METHOD name=Workflow.decorate qn=Workflow.decorate canonical=<ROOT>/fidelity.rb:Workflow.decorate#0 line=35 col=3 end=37:6 file=FILE:<ROOT>/fidelity.rb@1
+node METHOD name=Workflow.initialize qn=Workflow.initialize canonical=<ROOT>/fidelity.rb:Workflow.initialize#0 line=24 col=3 end=27:6 file=FILE:<ROOT>/fidelity.rb@1
+node METHOD name=Workflow.run qn=Workflow.run canonical=<ROOT>/fidelity.rb:Workflow.run#0 line=29 col=3 end=33:6 file=FILE:<ROOT>/fidelity.rb@1
+node METHOD name=orchestrate_ruby qn=orchestrate_ruby canonical=<ROOT>/fidelity.rb:orchestrate_ruby#0 line=40 col=1 end=42:4 file=FILE:<ROOT>/fidelity.rb@1
+node MODULE name="logger" qn="logger" canonical=<ROOT>/fidelity.rb:"logger"#0 line=1 col=9 end=1:17 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=decorate qn=decorate canonical=<ROOT>/fidelity.rb:decorate#0 line=32 col=5 end=32:13 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=name qn=name canonical=<ROOT>/fidelity.rb:name#0 line=5 col=11 end=5:15 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=name qn=name canonical=<ROOT>/fidelity.rb:name#1 line=11 col=16 end=11:20 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=name qn=name canonical=<ROOT>/fidelity.rb:name#2 line=17 col=11 end=17:15 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=name qn=name canonical=<ROOT>/fidelity.rb:name#3 line=36 col=11 end=36:15 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=new qn=new canonical=<ROOT>/fidelity.rb:new#0 line=21 col=16 end=21:19 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=new qn=new canonical=<ROOT>/fidelity.rb:new#1 line=41 col=63 end=41:66 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=notify qn=notify canonical=<ROOT>/fidelity.rb:notify#0 line=30 col=15 end=30:21 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=puts qn=puts canonical=<ROOT>/fidelity.rb:puts#0 line=11 col=5 end=11:9 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=require qn=require canonical=<ROOT>/fidelity.rb:require#0 line=1 col=1 end=1:8 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=run qn=run canonical=<ROOT>/fidelity.rb:run#0 line=41 col=53 end=41:56 file=FILE:<ROOT>/fidelity.rb@1
+node UNKNOWN name=save qn=save canonical=<ROOT>/fidelity.rb:save#0 line=31 col=17 end=31:21 file=FILE:<ROOT>/fidelity.rb@1
+edge CALL src=METHOD:ConsoleNotifier.notify@10 dst=UNKNOWN:name@11 resolved_src=METHOD:ConsoleNotifier.notify@10 resolved_dst=- line=11 confidence=- certainty=- callsite=h0:11:16:h1|syntax:ruby-member-call candidates=[]
+edge CALL src=METHOD:ConsoleNotifier.notify@10 dst=UNKNOWN:puts@11 resolved_src=METHOD:ConsoleNotifier.notify@10 resolved_dst=- line=11 confidence=- certainty=- callsite=h0:11:5:h2 candidates=[]
+edge CALL src=METHOD:Notifier.notify@4 dst=UNKNOWN:name@5 resolved_src=METHOD:Notifier.notify@4 resolved_dst=- line=5 confidence=- certainty=- callsite=h0:5:11:h3|syntax:ruby-member-call candidates=[]
+edge CALL src=METHOD:Repository.save@16 dst=UNKNOWN:name@17 resolved_src=METHOD:Repository.save@16 resolved_dst=- line=17 confidence=- certainty=- callsite=h0:17:11:h4|syntax:ruby-member-call candidates=[]
+edge CALL src=METHOD:Workflow.decorate@35 dst=UNKNOWN:name@36 resolved_src=METHOD:Workflow.decorate@35 resolved_dst=- line=36 confidence=- certainty=- callsite=h0:36:11:h5|syntax:ruby-member-call candidates=[]
+edge CALL src=METHOD:Workflow.run@29 dst=UNKNOWN:decorate@32 resolved_src=METHOD:Workflow.run@29 resolved_dst=METHOD:Workflow.decorate@35 line=32 confidence=0.9500 certainty=Certain callsite=h0:32:5:h6 candidates=[]
+edge CALL src=METHOD:Workflow.run@29 dst=UNKNOWN:notify@30 resolved_src=METHOD:Workflow.run@29 resolved_dst=- line=30 confidence=- certainty=- callsite=h0:30:15:h7|syntax:ruby-member-call candidates=[]
+edge CALL src=METHOD:Workflow.run@29 dst=UNKNOWN:save@31 resolved_src=METHOD:Workflow.run@29 resolved_dst=- line=31 confidence=- certainty=- callsite=h0:31:17:h8|syntax:ruby-member-call candidates=[]
+edge CALL src=METHOD:orchestrate_ruby@40 dst=METHOD:Workflow.run@29 resolved_src=METHOD:orchestrate_ruby@40 resolved_dst=METHOD:Workflow.run@29 line=41 confidence=1.0000 certainty=Certain callsite=h0:41:1:h9 candidates=[]
+edge CALL src=METHOD:orchestrate_ruby@40 dst=UNKNOWN:new@41 resolved_src=METHOD:orchestrate_ruby@40 resolved_dst=- line=41 confidence=- certainty=- callsite=h0:41:12:h10 candidates=[]
+edge CALL src=METHOD:orchestrate_ruby@40 dst=UNKNOWN:new@41 resolved_src=METHOD:orchestrate_ruby@40 resolved_dst=- line=41 confidence=- certainty=- callsite=h0:41:32:h10 candidates=[]
+edge CALL src=METHOD:orchestrate_ruby@40 dst=UNKNOWN:new@41 resolved_src=METHOD:orchestrate_ruby@40 resolved_dst=- line=41 confidence=- certainty=- callsite=h0:41:48:h10 candidates=[]
+edge CALL src=METHOD:orchestrate_ruby@40 dst=UNKNOWN:new@41 resolved_src=METHOD:orchestrate_ruby@40 resolved_dst=- line=41 confidence=- certainty=- callsite=h0:41:63:h10 candidates=[]
+edge IMPORT src=MODULE:"logger"@1 dst=MODULE:"logger"@1 resolved_src=MODULE:"logger"@1 resolved_dst=- line=1 confidence=- certainty=- callsite=- candidates=[]
+edge INHERITANCE src=CLASS:ConsoleNotifier@9 dst=CLASS:Notifier@3 resolved_src=- resolved_dst=- line=0 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:ConsoleNotifier@9 dst=METHOD:ConsoleNotifier.notify@10 resolved_src=- resolved_dst=- line=10 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Notifier@3 dst=METHOD:Notifier.notify@4 resolved_src=- resolved_dst=- line=4 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Repository@15 dst=METHOD:Repository.save@16 resolved_src=- resolved_dst=- line=16 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Workflow@23 dst=METHOD:Workflow.decorate@35 resolved_src=- resolved_dst=- line=35 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Workflow@23 dst=METHOD:Workflow.initialize@24 resolved_src=- resolved_dst=- line=24 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Workflow@23 dst=METHOD:Workflow.run@29 resolved_src=- resolved_dst=- line=29 confidence=- certainty=Certain callsite=- candidates=[]
